+++ components/Header.tsx
@@ -49,7 +49,7 @@ export default function Header() {
                             key={item}
                             className="text-gray-300 hover:bg-gray-700 hover:text-white px-3 py-2 rounded-md text-sm font-medium"
                           >
-                            <Link href="/users/search">{item}</Link>
+                            <Link href="/users">{item}</Link>
                           </div>
                         )}
                         {itemIdx === 2 && (
